added signOut and decentralized code
